requirements.txt
pandas==1.1.4
tensorflow-metadata==0.24